Merging counts for keywords from new+ old dataset
Merging both appearances of keywords for the new dataset and the old one as well. We look only from 2014 to 2024. 
Trained polynomial regression model of order 2 to predict for 2025 and 2026. 
Tested other models as well, but poly order 2 is the best. 
Added ridge regression to it to avoid overfitting because the dataset is small.